Treat semicolons as a delimiter outside of groups.
This patch modifies addressparser to treat semicolons like commas when used
outside of a group.  I believe this may have been the original intent of
addressparser since there is explicit logic in addressparser.parse that already
treats semicolons identically to commas.

Context: In the Firefox OS email app we allow the user to use the semicolon
character to act as a delimiter in the UI for consistency with other email
client UIs and because there is no valid use for the semicolon apart from a
delimiter in that context.  The app currently depends on addressparser to treat
the semicolon as a delimiter identical to a comma when used outside of a group,
but addressparser does not.  I think it makes more sense to address this in
addressparser than in our UI in this situation, but we can also do that.

(Note that this was a regression in our UI, but it was probably due to the
change to have our UI use addressparser directly rather than the very naive
approach we were using previously that just looked for commas and spaces and
semicolons.)